app/requirements.txt
#need to lock the versions
kafka-python
pandas
requests
s3fs
awscli
datetime